Revamp Command folder + Psh setup and deploy command
