+++ src/Command/Docker/Start.php
@@ -4,7 +4,7 @@
  * @license   For full copyright and license information view LICENSE file distributed with this source code.
  */
 
-namespace eZ\Launchpad\Command;
+namespace eZ\Launchpad\Command\Docker;
 
 use eZ\Launchpad\Core\DockerCommand;
 use Symfony\Component\Console\Input\InputArgument;
@@ -12,9 +12,9 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Class Logs.
+ * Class Start.
  */
-class Logs extends DockerCommand
+class Start extends DockerCommand
 {
     /**
      * {@inheritdoc}
@@ -22,9 +22,9 @@ class Logs extends DockerCommand
     protected function configure()
     {
         parent::configure();
-        $this->setName('docker:logs')->setDescription('Display the logs.');
-        $this->addArgument('service', InputArgument::OPTIONAL, 'Service to log.', '');
-        $this->setAliases(['logs', 'log']);
+        $this->setName('docker:start')->setDescription('Start all the services (or just one).');
+        $this->addArgument('service', InputArgument::OPTIONAL, 'Service to start', '');
+        $this->setAliases(['start']);
     }
 
     /**
@@ -32,6 +32,6 @@ class Logs extends DockerCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $this->dockerClient->logs(['-f', '--tail=100'], $input->getArgument('service'));
+        $this->dockerClient->start($input->getArgument('service'));
     }
 }
