+++ src/Command/Docker/Status.php
@@ -4,17 +4,18 @@
  * @license   For full copyright and license information view LICENSE file distributed with this source code.
  */
 
-namespace eZ\Launchpad\Command;
+namespace eZ\Launchpad\Command\Docker;
 
 use eZ\Launchpad\Core\DockerCommand;
 use eZ\Launchpad\Core\ProjectStatusDumper;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Class Create.
+ * Class Status.
  */
-class Create extends DockerCommand
+class Status extends DockerCommand
 {
     /**
      * @var ProjectStatusDumper
@@ -38,8 +39,14 @@ class Create extends DockerCommand
     protected function configure()
     {
         parent::configure();
-        $this->setName('docker:create')->setDescription('Create all the services.');
-        $this->setAliases(['create']);
+        $this->setName('docker:status')->setDescription('Obtaining the project information.');
+        $this->setAliases(['docker:ps', 'docker:info', 'ps', 'info']);
+        $this->addArgument(
+            'options',
+            InputArgument::OPTIONAL,
+            'n: Docker Network, c: Docker Compose, s: Service Access',
+            'ncs'
+        );
     }
 
     /**
@@ -57,20 +64,6 @@ class Create extends DockerCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $this->dockerClient->build(['--no-cache']);
-        $this->dockerClient->up(['-d']);
-
-        $this->taskExecutor->composerInstall();
-        $this->taskExecutor->eZCreate();
-        $this->taskExecutor->importData();
-
-        // if solr run the index
-        $compose = $this->projectConfiguration->getDockerCompose();
-        if ($compose->hasService('solr')) {
-            $this->taskExecutor->createCore();
-            $this->taskExecutor->indexSolr();
-        }
-
-        $this->projectStatusDumper->dump('ncsi');
+        $this->projectStatusDumper->dump($input->getArgument('options'));
     }
 }
